+++ src/dev/build.ts
@@ -13,15 +13,17 @@ import {
   ensureFile,
   join,
   SEPARATOR,
-  resolve,
-  normalize,
-  toFileUrl
+  toFileUrl,
 } from "../deps.ts";
 import { fileExists } from "../server/utils.ts";
 import type { GetRouterOptions, Handlers } from "../server/router.ts";
 import type { AppTemplateInterface } from "../server/ssr.ts";
+import type { Middleware } from "../deps.ts";
 import { getIslandsRegistered } from "./extract-islands.ts";
 import { resolvePath, getTailwind } from "./path.ts";
+// @ts-ignore lit is a npm package and Deno doesn't resolve the exported members
+import type { LitElement } from "lit";
+import type { RouteModule, LayoutModule, MiddlewareModule } from "../types.ts";
 
 const TEST_FILE_PATTERN = /[._]test\.(?:[tj]sx?|[mc][tj]s)$/;
 
@@ -30,13 +32,17 @@ export type BuildRoute = {
   path: string;
   relativeFilePath: string;
   absoluteFilePath: string;
-  routeModule?: { default: unknown; handler?: Handlers };
+  routeModule?: RouteModule;
   tagName: string;
   jsAssetContent: esbuild.OutputFile | undefined;
   jsAssetPath: string | undefined;
   cssAssetContent: string | undefined;
   cssAssetPath: string | undefined;
   islands: string[] | undefined;
+  middlewares: MiddlewareModule[] | [];
+  middlewarePaths: string[] | [];
+  layouts: LayoutModule[] | [];
+  layoutPaths: string[] | [];
 };
 
 const encoder = new TextEncoder();
@@ -82,7 +88,7 @@ async function buildJS(path: string, { devMode }: { devMode?: boolean }) {
   ].join("\n");
 
   const tempDirPath = await Deno.makeTempDir();
-  const entrypointPath = join(tempDirPath, '/entrypoint.js');
+  const entrypointPath = join(tempDirPath, "/entrypoint.js");
   await Deno.writeTextFile(entrypointPath, entryPoint, {});
 
   const esbuildResult = await esbuild.build({
@@ -230,12 +236,19 @@ export async function getRoutes({
 
   const ignoreFilePattern = TEST_FILE_PATTERN;
   const routes: BuildRoute[] = [];
+  const allMiddlewareFiles = await getMiddlewareFiles();
+  const allLayoutFiles = await getLayoutFiles();
 
   for await (const entry of walk("./routes", {
     includeDirs: false,
     includeSymlinks: false,
     exts: ["ts", "js"],
-    skip: [ignoreFilePattern, new RegExp("/_app.(js|ts)$")],
+    skip: [
+      ignoreFilePattern,
+      new RegExp("/_app.(js|ts)$"),
+      new RegExp("/_middleware.(js|ts)$"),
+      new RegExp("/_layout.(js|ts)$"),
+    ],
   })) {
     const parsed = parse(entry.path);
     let path = parsed.dir.replace("routes", "") + "/" + parsed.name;
@@ -276,23 +289,185 @@ export async function getRoutes({
         ? `/_limette/css/tailwind-${id}.css`
         : undefined;
 
+    // For buildAssets, we set the middlewares as an array of paths, otherwise an array of modules
+    const middlewares = devMode
+      ? ((await getMiddlewaresForRoute({
+          filePath: entry.path,
+          allMiddlewareFiles: allMiddlewareFiles,
+          loadFs: loadFs,
+          valueType: "module",
+        })) as MiddlewareModule[])
+      : [];
+
+    const middlewarePaths =
+      !devMode && buildAssets
+        ? ((await getMiddlewaresForRoute({
+            filePath: entry.path,
+            allMiddlewareFiles: allMiddlewareFiles,
+            loadFs: loadFs,
+            valueType: "absolutePath",
+          })) as string[])
+        : [];
+
+    const layouts = (await getLayoutsForRoute({
+      filePath: entry.path,
+      allLayoutFiles: allLayoutFiles,
+      loadFs: loadFs,
+      valueType: "module",
+    })) as LayoutModule[];
+
+    const layoutPaths = (await getLayoutsForRoute({
+      filePath: entry.path,
+      allLayoutFiles: allLayoutFiles,
+      loadFs: loadFs,
+      valueType: "absolutePath",
+    })) as string[];
+
     const route: BuildRoute = {
       id,
       path,
-      relativeFilePath:`.${SEPARATOR}${entry.path}`,
+      relativeFilePath: `.${SEPARATOR}${entry.path}`,
       absoluteFilePath,
-      routeModule: (await loadFs?.(entry.path)) as { default: unknown },
+      routeModule: (await loadFs?.(entry.path)) as RouteModule,
       tagName,
       jsAssetContent,
       jsAssetPath,
       cssAssetContent,
       cssAssetPath,
       islands,
+      middlewares: middlewares,
+      middlewarePaths: middlewarePaths,
+      layouts: layouts,
+      layoutPaths: layoutPaths,
     };
     routes.push(route);
   }
 
   return routes;
+}
+
+// Method to get all _middleware files
+async function getMiddlewareFiles(): Promise<Map<string, string>> {
+  const middlewareFiles = new Map();
+  for await (const entry of walk("./routes", {
+    includeDirs: false,
+    includeSymlinks: false,
+    exts: ["ts", "js"],
+    match: [new RegExp("/_middleware.(ts|js)$")],
+  })) {
+    const parsed = parse(entry.path);
+    middlewareFiles.set(parsed.dir, entry.path);
+  }
+  return middlewareFiles;
+}
+
+// Method to find middleware files in the tree of a given file path
+async function getMiddlewaresForRoute({
+  filePath,
+  allMiddlewareFiles,
+  loadFs,
+  valueType,
+}: {
+  filePath: string;
+  allMiddlewareFiles: Map<string, string>;
+  loadFs?: (path: string) => Promise<unknown>;
+  valueType: "module" | "relativePath" | "absolutePath";
+}): Promise<MiddlewareModule[] | string[] | []> {
+  if (allMiddlewareFiles.size === 0) return [];
+
+  // Remove file name from path
+  const pathSegments = filePath.split("/").slice(0, -1); // ["foo", "bar"]
+
+  // Build directory paths from root to the file's directory using a for...of loop
+  const directoriesToCheck: string[] = [];
+  let currentDir = "";
+
+  for (const segment of pathSegments) {
+    currentDir = join(currentDir, segment);
+    directoriesToCheck.push(currentDir);
+  }
+
+  const middlewareFilesForRoute = directoriesToCheck
+    .filter((dir) => allMiddlewareFiles.has(dir))
+    .map((dir) => allMiddlewareFiles.get(dir) ?? "")
+    .filter(Boolean);
+
+  // Return array of relative paths
+  if (valueType === "relativePath") {
+    return middlewareFilesForRoute;
+  }
+
+  // Return array of absolute paths
+  if (valueType === "absolutePath") {
+    return middlewareFilesForRoute.map((path) => join(Deno.cwd(), path));
+  }
+
+  return await Promise.all(
+    middlewareFilesForRoute.map(
+      (file) => loadFs?.(file) as unknown as MiddlewareModule
+    )
+  );
+}
+
+// Method to get all _layout files
+async function getLayoutFiles(): Promise<Map<string, string>> {
+  const layoutFiles = new Map();
+  for await (const entry of walk("./routes", {
+    includeDirs: false,
+    includeSymlinks: false,
+    exts: ["ts", "js"],
+    match: [new RegExp("/_layout.(ts|js)$")],
+  })) {
+    const parsed = parse(entry.path);
+    layoutFiles.set(parsed.dir, entry.path);
+  }
+  return layoutFiles;
+}
+
+// Method to find middleware files in the tree of a given file path
+async function getLayoutsForRoute({
+  filePath,
+  allLayoutFiles,
+  loadFs,
+  valueType,
+}: {
+  filePath: string;
+  allLayoutFiles: Map<string, string>;
+  loadFs?: (path: string) => Promise<unknown>;
+  valueType: "module" | "relativePath" | "absolutePath";
+}): Promise<LayoutModule[] | string[] | []> {
+  if (allLayoutFiles.size === 0) return [];
+
+  // Remove file name from path
+  const pathSegments = filePath.split("/").slice(0, -1); // ["foo", "bar"]
+
+  // Build directory paths from root to the file's directory using a for...of loop
+  const directoriesToCheck: string[] = [];
+  let currentDir = "";
+
+  for (const segment of pathSegments) {
+    currentDir = join(currentDir, segment);
+    directoriesToCheck.push(currentDir);
+  }
+
+  const layoutFilesForRoute = directoriesToCheck
+    .filter((dir) => allLayoutFiles.has(dir))
+    .map((dir) => allLayoutFiles.get(dir) ?? "")
+    .filter(Boolean);
+
+  // Return array of relative paths
+  if (valueType === "relativePath") {
+    return layoutFilesForRoute;
+  }
+
+  // Return array of absolute paths
+  if (valueType === "absolutePath") {
+    return layoutFilesForRoute.map((path) => join(Deno.cwd(), path));
+  }
+
+  return await Promise.all(
+    layoutFilesForRoute.map((file) => loadFs?.(file) as unknown as LayoutModule)
+  );
 }
 
 export async function getAppTemplate({ loadFs }: GetRouterOptions) {
@@ -361,6 +536,10 @@ export async function build() {
   let routeIndex = 0;
   let routesImportsString = ``;
   let routesArrayString = ``;
+  let middlewareImportsString = ``;
+  const importedMiddlewares = new Map();
+  let layoutImportsString = `\n`;
+  const importedLayouts = new Map();
 
   for (const route of routes) {
     if (route.jsAssetContent?.contents) {
@@ -377,8 +556,30 @@ export async function build() {
     }
 
     // Generate static routes file
-    routesImportsString += `import * as route${routeIndex} from "${toFileUrl(route.absoluteFilePath)}";
+    routesImportsString += `import * as route${routeIndex} from "${toFileUrl(
+      route.absoluteFilePath
+    )}";
 `;
+
+    middlewareImportsString += route.middlewarePaths
+      .filter((path) => !importedMiddlewares.has(path))
+      .map((path) => {
+        const middlewareModuleName = `middleware${importedMiddlewares.size}`;
+        importedMiddlewares.set(path, middlewareModuleName);
+        return `\nimport * as ${middlewareModuleName} from "${toFileUrl(
+          path
+        )}";`;
+      })
+      .join("");
+
+    layoutImportsString += route.layoutPaths
+      .filter((path) => !importedLayouts.has(path))
+      .map((path) => {
+        const layoutModuleName = `layout${importedLayouts.size}`;
+        importedLayouts.set(path, layoutModuleName);
+        return `\nimport * as ${layoutModuleName} from "${toFileUrl(path)}";`;
+      })
+      .join("");
 
     routesArrayString += `
   {
@@ -392,16 +593,25 @@ export async function build() {
     cssAssetPath: ${
       route.cssAssetPath ? `"${route.cssAssetPath}"` : `undefined`
     },
-    islands: ${JSON.stringify(route.islands)}
+    islands: ${JSON.stringify(route.islands)},
+    middlewares: [${route.middlewarePaths
+      .map((path) => importedMiddlewares.get(path))
+      .join()}],
+    layouts: [${route.layoutPaths
+      .map((path) => importedLayouts.get(path))
+      .join()}]
   },
 `;
     routeIndex++;
   }
 
-  routesArrayString = `export const routes = [${routesArrayString}];`;
+  routesArrayString = `\n\nexport const routes = [${routesArrayString}];`;
 
   await Deno.writeTextFile(
     "./_limette/routes.js",
-    routesImportsString + routesArrayString
+    routesImportsString +
+      middlewareImportsString +
+      layoutImportsString +
+      routesArrayString
   );
 }